requirements.txt
python-docx
python-decouple